// src/common/enum.ts
export enum IWindowList {
  PET_WINDOW = 'PET_WINDOW',
  PET_DETAIL_WINDOW = 'PET_DETAIL_WINDOW',
  PET_CHAT_WINDOW = 'PET_CHAT_WINDOW',
  PET_SETTING_WINDOW = 'PET_SETTING_WINDOW',
}

export enum DBList {
  Message_DB = 'Message_DB',
  Config_DB = 'Config_DB',
  Prompt_DB = 'Prompt_DB',
  History_DB = 'History_DB',
}